refactor(LottoService): 수익률 계산 시 오버플로 방지를 위해 totalPrize를 long으로 변경
- totalPrize 타입을 int에서 long으로 변경해 대규모 상금 계산 시 안정성 확보
- 1등 상금이 여러 개 당첨되는 경우에도 오버플로 발생 방지
- 로직 변경 없이 계산 안정성 강화

+++ src/main/java/lotto/domain/service/LottoService.java
@@ -48,8 +48,8 @@ public class LottoService {
     }
 
     public double calculateProfitPurchaseAmount(PurchaseAmount purchaseAmount, Map<Rank, Long> rankCounts) {
-        int totalPrize = rankCounts.entrySet().stream()
-                .mapToInt(entry -> entry.getKey().getPrize() * entry.getValue().intValue())
+        long totalPrize = rankCounts.entrySet().stream()
+                .mapToLong(entry -> (long) entry.getKey().getPrize() * entry.getValue().intValue())
                 .sum();
 
         return (double) totalPrize / purchaseAmount.getPurchaseAmount() * PERCENT;
